moved packages around and solved some issues and reopened one kinda
