+++ src/physicalOutput/TalonMotor.java
@@ -1,24 +1,23 @@
-package motors;
-import edu.wpi.first.wpilibj.Spark;
+package physicalOutput;
+import edu.wpi.first.wpilibj.Talon;
 
-//more leinad's stuff
 
 /**
  * 
  */
-public class SparkMotor extends IMotor {
+public class TalonMotor extends IMotor {
 
     /**
      * Default constructor
      */
-    public SparkMotor(int x) {
-    	part = new Spark(x);
+    public TalonMotor(int x) {
+    	part = new Talon(x);
     }
 
     /**
      * 
      */
-    private Spark part;
+    private Talon part;
 
     /**
      * Sets the motor pwm value
